restructuring graph interface wip wip wip
